+++ tests/03_run/cmd/test_run_cmd.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,7 +11,7 @@ class Test:
         os.chdir(Path(__file__).parent)
 
     def test_run_cmd(self, capsys):
-        cmd2 = ["-w", "50", "-m", "exec", "solver.yaml", "cases.tio"]
+        cmd2 = ["-w", "50", "-m", "run", "solver.yaml", "cases.tio"]
         Compare.list(capsys, "out1", cmd2)