+++ tests/03_run/fup/test_run_fup.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,23 +11,23 @@ class Test:
         os.chdir(Path(__file__).parent)
 
     def test_run_mixed_side(self, capsys):
-        cmd = "-w 80 -m exec solver.cpp cases.tio -as"
+        cmd = "-w 80 -m run solver.cpp cases.tio -as"
         Compare.text(capsys, "out1", cmd)
 
     def test_run_side_0(self, capsys):
-        cmd = "-w 80 -m exec solver.py cases.tio -d"
+        cmd = "-w 80 -m run solver.py cases.tio -d"
         Compare.text(capsys, "out3", cmd)
         
     def test_run_side_1(self, capsys):
-        cmd = "-w 80 -m exec solver.py cases.tio -s"
+        cmd = "-w 80 -m run solver.py cases.tio -s"
         Compare.text(capsys, "out4", cmd)
 
     def test_run_side_2(self, capsys):
-        cmd = "-w 80 -m exec solver.py cases.tio -ad"
+        cmd = "-w 80 -m run solver.py cases.tio -ad"
         Compare.text(capsys, "out5", cmd)
 
     def test_run_side_5(self, capsys):
-        cmd = "-w 80 -m exec solver.cpp cases.tio -ad"
+        cmd = "-w 80 -m run solver.cpp cases.tio -ad"
         Compare.text(capsys, "out2", cmd)
 
 if __name__ == '__main__':