+++ tests/03_run/lang/test_run_yaml.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,7 +11,7 @@ class Test:
         os.chdir(Path(__file__).parent)
 
     def test_run_mixed_side(self, capsys):
-        cmd = "-w 80 -m exec solver.yaml cases.tio -as"
+        cmd = "-w 80 -m run solver.yaml cases.tio -as"
         Compare.text(capsys, "out1", cmd)