+++ tests/03_run/overflow/test_run_overflow.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,19 +11,19 @@ class Test:
         os.chdir(Path(__file__).parent)
 
     # def test_run_mixed_1(self, capsys):
-    #     Compare.text(capsys, "out1", "-w 80 -m exec runtime.cpp cases.tio -ad")
+    #     Compare.text(capsys, "out1", "-w 80 -m run runtime.cpp cases.tio -ad")
 
     # def test_run_mixed_2(self, capsys):
-    #     Compare.text(capsys, "-w 80 -m exec runtime.cpp", "out2")
+    #     Compare.text(capsys, "-w 80 -m run runtime.cpp", "out2")
 
     def test_run_mixed_3(self, capsys):
-        Compare.text(capsys, "out2", "-w 80 -m exec exception.py -d")
+        Compare.text(capsys, "out2", "-w 80 -m run exception.py -d")
 
     # def test_run_mixed_4(self, capsys):
-    #     Compare.text(capsys, "out3", "-w 80 -m exec exception.cpp cases.tio -d")
+    #     Compare.text(capsys, "out3", "-w 80 -m run exception.cpp cases.tio -d")
 
     # def test_run_mixed_5(self, capsys):
-    #     Compare.text(capsys, "out4", "-w 80 -m exec exception.py cases.tio -s")
+    #     Compare.text(capsys, "out4", "-w 80 -m run exception.py cases.tio -s")
 
 if __name__ == '__main__':
     unittest.main()