+++ tests/03_run/poo/test_run_poo.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,13 +11,13 @@ class Test:
         os.chdir(Path(__file__).parent)
             
     def test_run_mixed_1(self, capsys):
-        Compare.text(capsys, "out1", "-w 80 -m exec draft.ts cases.tio -s")
+        Compare.text(capsys, "out1", "-w 80 -m run draft.ts cases.tio -s")
 
     def test_run_mixed_2(self, capsys):
-        Compare.text(capsys, "out2", "-w 80 -m exec solver.cpp cases.tio -s")
+        Compare.text(capsys, "out2", "-w 80 -m run solver.cpp cases.tio -s")
 
     def test_run_mixed_3(self, capsys):
-        Compare.text(capsys, "out3", "-w 80 -m exec draft.ts cases.tio -d")
+        Compare.text(capsys, "out3", "-w 80 -m run draft.ts cases.tio -d")