+++ tests/03_run/sql/test_run_sql.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,11 +11,11 @@ class Test:
         os.chdir(Path(__file__).parent)
 
     def test_run_1(self, capsys):
-        cmd = ["-m", "exec", "solver.yaml", "cases.tio"]
+        cmd = ["-m", "run", "solver.yaml", "cases.tio"]
         Compare.list(capsys, "out1", cmd)
 
     def test_run_2(self, capsys):
-        cmd = ["-m", "exec", "solver.yaml", "wrong.tio", "-s"]
+        cmd = ["-m", "run", "solver.yaml", "wrong.tio", "-s"]
         Compare.list(capsys, "out2", cmd)