+++ tests/04_diff/test_diff.py
@@ -2,7 +2,7 @@ import unittest
 import os
 from pathlib import Path
 
-from tko.__main__ import exec, Parser #type: ignore
+from tko.__main__ import execute, Parser #type: ignore
 from tko.util.compare import Compare # type: ignore
 
 class Test:
@@ -11,11 +11,11 @@ class Test:
         os.chdir(Path(__file__).parent)
                 
     def test_run_mixed_1(self, capsys):
-        Compare.text(capsys, "out1", "-w 80 -m exec cases.tio solver.py -s")
+        Compare.text(capsys, "out1", "-w 80 -m run cases.tio solver.py -s")
 
 
     def test_run_mixed_2(self, capsys):
-        Compare.text(capsys, "out2", "-w 80 -m exec cases.tio solver.py -d")
+        Compare.text(capsys, "out2", "-w 80 -m run cases.tio solver.py -d")
 
 if __name__ == '__main__':
     unittest.main()